+++ noodles-cram/src/writer.rs
@@ -272,7 +272,7 @@ where
     fn write_alignment_record(
         &mut self,
         header: &sam::Header,
-        record: &sam::alignment::RecordBuf,
+        record: &dyn sam::alignment::Record,
     ) -> io::Result<()> {
         let r = Record::try_from_alignment_record(header, record)?;
         self.write_record(header, r)