+++ noodles-sam/src/alignment_writer.rs
@@ -1,6 +1,6 @@
 use std::io;
 
-use super::{alignment::RecordBuf, Header};
+use super::{alignment::Record, Header};
 
 /// An alignment format writer.
 ///
@@ -10,7 +10,7 @@ pub trait AlignmentWriter {
     fn write_alignment_header(&mut self, header: &Header) -> io::Result<()>;
 
     /// Writes an alignment record.
-    fn write_alignment_record(&mut self, header: &Header, record: &RecordBuf) -> io::Result<()>;
+    fn write_alignment_record(&mut self, header: &Header, record: &dyn Record) -> io::Result<()>;
 
     /// Shuts down an alignment format writer.
     fn finish(&mut self, header: &Header) -> io::Result<()>;