+++ noodles-sam/CHANGELOG.md
@@ -31,6 +31,9 @@
   * sam/alignment/record_buf: Change quality scores to raw scores
     (`record_buf::QualityScores`).
 
+  * sam/alignment_writer: Change `AlignmentWriter::write_record` to accept
+    `&dyn crate::alignment::Record`.
+
   * sam/lazy/record/data/field: Replace `Value` with
     `crate::alignment::record::data::field::Value`.